+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Suggest
 - Added support for Fakespot suggestions.
+- Removed the `SuggestIngestionConstraints::max_suggestions` field.  No consumers were using this.
 
 ## 🦊 What's Changed 🦊
 
@@ -24,7 +25,6 @@
 
 ### Suggest
 - Removed the deprecated `remote_settings_config` method.  No consumers were using this.
-- Removed the `SuggestIngestionConstraints::max_suggestions` field.  No consumers were using this.
 
 ## ✨ What's New ✨